requirements.txt
numpy==1.21.5
pandas==1.3.5
opencv-python==4.5.5.64
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.3
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
